tidied up timers
